requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openenv-core>=0.1.0
numpy>=1.26.0
matplotlib>=3.8.0
